combat: 引入 AttackParams，重构潜行攻击接口并修复 Clippy 警告
- 在 combat/lib.rs 引入 AttackParams 参数包，重构 perform_attack_with_ambush 以减少长参数数量
- 更新 CombatManager 以使用 AttackParams，移除本地重复定义，调整测试
- 修复 combatant.rs 中不必要的类型转换与 attack_distance 中的 unnecessary_cast
- 修复 effect.rs 中的 manual_is_multiple_of，采用 is_multiple_of 简化逻辑
- 修复 enemy.rs 中的 unnecessary_min_or_max，使用 saturating_sub 防溢出
- 在 vision.rs 为多参数函数添加 #[allow(clippy::too_many_arguments)]
- 运行 cargo clippy -p combat -- -D warnings，通过零警告

items: 规范 if-let 复合条件的换行格式，保持 Rust 2024 风格一致

+++ src/combat/src/combatant.rs
@@ -89,7 +89,7 @@ impl Combatant for Enemy {
     }
 
     fn attack_power(&self) -> u32 {
-        self.attack + self.weapon.as_ref().map_or(0, |w| w.damage_bonus() as u32)
+        self.attack + self.weapon.as_ref().map_or(0, |w| w.damage_bonus())
     }
 
     fn defense(&self) -> u32 {
@@ -157,7 +157,7 @@ impl Combatant for Enemy {
     fn attack_distance(&self) -> u32 {
         self.weapon
             .as_ref()
-            .map_or(self.attack_range, |w| w.range() as u32)
+            .map_or(self.attack_range, |w| w.range())
     }
 
     fn take_damage(&mut self, amount: u32) -> bool {
